+++ drivers/gpu/drm/apple/Makefile
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: GPL-2.0-only OR MIT
 
 appledrm-y := apple_drv.o
-apple_dcp-y := dcp.o parser.o
+apple_dcp-y := dcp.o iomfb.o parser.o
 apple_piodma-y := dummy-piodma.o
 
 obj-$(CONFIG_DRM_APPLE) += appledrm.o
