+++ drivers/gpu/drm/apple/dcp.h
@@ -5,6 +5,9 @@
 #define __APPLE_DCP_H__
 
 #include <drm/drm_atomic.h>
+#include <drm/drm_fourcc.h>
+
+#include "dcp-internal.h"
 #include "parser.h"
 
 struct apple_crtc {
@@ -39,8 +42,16 @@ void dcp_link(struct platform_device *pdev, struct apple_crtc *apple,
 void dcp_flush(struct drm_crtc *crtc, struct drm_atomic_state *state);
 bool dcp_is_initialized(struct platform_device *pdev);
 void apple_crtc_vblank(struct apple_crtc *apple);
+void dcp_drm_crtc_vblank(struct apple_crtc *crtc);
 int dcp_get_modes(struct drm_connector *connector);
 int dcp_mode_valid(struct drm_connector *connector,
 		   struct drm_display_mode *mode);
+void dcp_set_dimensions(struct apple_dcp *dcp);
+
+
+int iomfb_start_rtkit(struct apple_dcp *dcp);
+void iomfb_shutdown(struct apple_dcp *dcp);
+/* rtkit message handler for IOMFB messages */
+void iomfb_recv_msg(struct apple_dcp *dcp, u64 message);
 
 #endif
